更名

+++ Assets/Scripts/KouXiaGu/Terrain3D/Bake/Renderer.NormalMap.cs
@@ -109,7 +109,7 @@ namespace KouXiaGu.Terrain3D
             public Texture2D GetTexture(RenderTexture rt)
             {
                 RenderTexture.active = rt;
-                Texture2D normalMap = new Texture2D(Parameter.HeightMapWidth, Parameter.HeightMapHeight, TextureFormat.ARGB32, false);
+                Texture2D normalMap = new Texture2D((int)Parameter.HeightMapWidth, (int)Parameter.HeightMapHeight, TextureFormat.ARGB32, false);
                 normalMap.ReadPixels(Parameter.HeightReadPixel, 0, 0, false);
                 normalMap.wrapMode = TextureWrapMode.Clamp;
                 normalMap.Apply();
